remove duplicate method
