+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/DataSourceV2Strategy.scala
@@ -344,8 +344,7 @@ class DataSourceV2Strategy(session: SparkSession) extends Strategy with Predicat
       WriteDeltaExec(planLater(query), refreshCache(r), projections, write) :: Nil
 
     case MergeRows(isSourceRowPresent, isTargetRowPresent, matchedInstructions,
-        notMatchedInstructions, notMatchedBySourceInstructions, checkCardinality, output,
-        child) =>
+        notMatchedInstructions, notMatchedBySourceInstructions, checkCardinality, output, child) =>
       MergeRowsExec(isSourceRowPresent, isTargetRowPresent, matchedInstructions,
         notMatchedInstructions, notMatchedBySourceInstructions, checkCardinality,
         output, planLater(child)) :: Nil
